+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 #########
 
+
+2.0.0
+*****
+
+- changed class members from protected to private
+- cs fixes, added codestyle checks
+
+
 1.0.0
 *****